# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI документация для фулфилмент сервиса
  version: v1
servers:
- url: http://localhost:8080
  description: Url сервера
...
paths:
...
  /api/v1/fulfillment:
    get:
      tags:
      - Fulfillment API
      summary: Поиск фулфилментов
      description: Возвращает список фулфилментов пользователя с фильтром и пагинацией
        в отсортированном по дате виде (сначала новые)
      operationId: getAllFulfillments
      parameters:
      - name: search
        in: query
        description: "Строка для поиска фулфилментов по началу наименования, артик\
          ула или баркода товара"
        required: false
        schema:
          type: string
        example: костюм
      - name: page
        in: query
        description: Номер страницы (начинается с 0)
        required: false
        schema:
          type: integer
          minimum: 0
        example: 0
      - name: limit
        in: query
        description: Количество элементов на странице
        required: false
        schema:
          type: integer
          maximum: 30
          minimum: 1
        example: 5
      responses:
        "200":
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllFulfillmentsResponse"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T19:34:23.56Z
                status: 400
                error: Bad request
                message: The limit cannot exceed 30
                path: /api/v1/fulfillment
        "401":
          description: Пользователь не аутентифицирован через telegram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:28:24.68Z
                status: 401
                error: Telegram authentication failed
                message: Init data expired. Log in again via Telegram
                path: /api/v1/fulfillment
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 500
                error: Database error
                message: Couldn't get fulfillments
                path: /api/v1/fulfillment
      security:
      - TelegramAuth: []
    post:
      tags:
      - Fulfillment API
      summary: Создание фулфилмента
      description: Сохраняет новый фулфилмент и отправляет данные в гугл таблицу
      operationId: createFulfillment
      requestBody:
        description: Данные для создания фулфилмента
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FulfillmentRequest"
        required: true
      responses:
        "201":
          description: Успешное сохранение
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FulfillmentResponse"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T18:42:51.70Z
                status: 400
                error: Validation error
                message: Incorrect data was entered
                path: /api/v1/fulfillment
                details:
                  quantity: The number must be greater than 0
                  productId: The product id cannot be empty
        "401":
          description: Пользователь не аутентифицирован через telegram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:28:24.68Z
                status: 401
                error: Telegram authentication failed
                message: Init data expired. Log in again via Telegram
                path: /api/v1/fulfillment
        "403":
          description: Доступ к товару запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T18:50:50.54Z
                status: 403
                error: Access denied
                message: There is no access to this product
                path: /api/v1/fulfillment
        "404":
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T18:48:20.03Z
                status: 404
                error: No data found
                message: "Product with id: 100152101 not found"
                path: /api/v1/fulfillment
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 500
                error: Database error
                message: Couldn't save fulfillment
                path: /api/v1/fulfillment
        "502":
          description: Ошибка при работе с api Google Sheets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T11:13:43.69Z
                status: 502
                error: Google Sheets API error
                message: "Couldn't export to Google Sheets, so the fulfillment was\
                  \ not saved"
                path: /api/v1/fulfillment
      security:
      - TelegramAuth: []
...
  /api/v1/fulfillment/{fulfillmentId}:
    get:
      tags:
      - Fulfillment API
      summary: Получение фулфилмента
      description: Возвращает информацию о фулфилменте
      operationId: getFulfillment
      parameters:
      - name: fulfillmentId
        in: path
        description: Id фулфилмента
        required: true
        schema:
          type: integer
          format: int64
          example: 10000005
      responses:
        "200":
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FulfillmentResponse"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T19:18:03.118+00:00
                status: 400
                error: Bad Request
                path: /api/v1/fulfillment/abcd
        "401":
          description: Пользователь не аутентифицирован через telegram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:28:24.68Z
                status: 401
                error: Telegram authentication failed
                message: Init data expired. Log in again via Telegram
                path: /api/v1/fulfillment/10000005
        "403":
          description: Доступ к фулфилменту запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T19:21:10.74Z
                status: 403
                error: Access denied
                message: There is no access to this fulfillment
                path: /api/v1/fulfillment/10000005
        "404":
          description: Фулфилмент не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T19:21:45.21Z
                status: 404
                error: No data found
                message: "Fulfillment with id: 100000 not found"
                path: /api/v1/fulfillment/100000
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                timestamp: 2025-07-28T13:22:31.00Z
                status: 500
                error: Database error
                message: Couldn't get fulfillment
                path: /api/v1/fulfillment/10000005
      security:
      - TelegramAuth: []
...
components:
  schemas:
...
    AppErrorResponse:
      type: object
      description: Стандартный ответ ошибки
      properties:
        timestamp:
          type: string
          description: "Дата и время, когда произошла ошибка"
          example: 2025-07-28T11:51:39.39Z
        status:
          type: integer
          format: int32
          description: Статусный код HTTP
          example: 400
        error:
          type: string
          description: Заголовок ошибки
          example: Validation error
        message:
          type: string
          description: Описание ошибки
          example: Incorrect data was entered
        path:
          type: string
          description: "Путь запроса, где произошла ошибка"
          example: /api/v1/shipment
        details:
          type: object
          additionalProperties:
            type: string
          description: Дополнительные сведения при ошибках валидации
          example:
            quantity: The number must be greater than 0
            warehouseId: The warehouse id cannot be empty
            warehouseName: The warehouse name cannot be empty
...
    FulfillmentResponse:
      type: object
      description: Ответ с информацией о фулфилменте
      properties:
        id:
          type: integer
          format: int64
          description: Id фулфилмента
          example: 10006115
        date:
          type: string
          description: Дата и время создания фулфилмента
          example: 2025-07-28T19:08:58.22Z
        title:
          type: string
          description: Наименование товара
          example: Костюм шорты рубашка летний
        barcode:
          type: string
          description: Баркод товара
          example: 2037991665923
        article:
          type: string
          description: Артикул товара
          example: костюм с шортами 850 бежевый
        size:
          type: string
          description: Размер товара
          example: L
        color:
          type: string
          description: Цвет товара
          example: бежевый
        photoUrl:
          type: string
          description: Ссылка на первое фото товара на Wildberries
          example: https://basket-11.wbbasket.ru/vol1647/part164710/164710105/images/big/1.webp
        quantity:
          type: integer
          format: int32
          description: Количество товара
          example: 200
        taskDescription:
          type: string
          description: Техническое задание
          example: "Проверить на брак, упаковать в пупырчатую пленку"
    FulfillmentRequest:
      type: object
      description: Запрос на создание фулфилмента
      properties:
        productId:
          type: integer
          format: int64
          description: id товара
          example: 10000330
          minimum: 0
        quantity:
          type: integer
          format: int32
          description: Количество
          example: 200
        taskDescription:
          type: string
          description: Техническое задание
          example: "Проверить на брак, упаковать в пупырчатую пленку"
          minLength: 1
      required:
      - productId
      - quantity
      - taskDescription
...
    AllFulfillmentsResponse:
      type: object
      description: Ответ на получение фулфилментов
      properties:
        fulfillments:
          type: array
          description: Список фулфилментов
          items:
            $ref: "#/components/schemas/FulfillmentResponse"
        size:
          type: integer
          format: int32
          description: Количество фулфилментов в ответе
          example: 1
        hasNext:
          type: boolean
          description: Наличие фулфилментов на следующей странице
          example: false
...
  securitySchemes:
    TelegramAuth:
      type: apiKey
      description: Telegram WebApp initData
      name: Authorization
      in: header
    ApiSecretAuth:
      type: apiKey
      description: Секретный ключ API
      name: X-API-SECRET
      in: header